size_T replaced with int. CurrentUserRequest implementation on client side

+++ Server/DataProviders/idataprovider.h
@@ -10,7 +10,7 @@ class IDataProvider : public QObject
 public:
     explicit IDataProvider(QObject *parent = nullptr);
     virtual void        addToHistory(Message msg) = 0;
-    virtual HistoryData getHistory(size_t firstUserId, size_t secondUserId, size_t size) = 0;
+    virtual HistoryData getHistory(int firstUserId, int secondUserId, int size) = 0;
     virtual UserData    getUserData(LoginPackage data) = 0;
 signals:
